refactor storage-models
